redone votemanager (unstable)
