+++ src/engine/shared/datafile.cpp
@@ -10,8 +10,6 @@
 
 #include "uuid_manager.h"
 
-static const int DEBUG = 0;
-
 enum
 {
 	OFFSET_UUID_TYPE = 0x8000,
@@ -238,9 +236,9 @@ bool CDataFileReader::Open(class IStorage *pStorage, const char *pFilename, int 
 
 	dbg_msg("datafile", "loading done. datafile='%s'", pFilename);
 
-	if(DEBUG)
+	/*if(DEBUG)
 	{
-		/*
+		
 		for(int i = 0; i < m_pDataFile->data.num_raw_data; i++)
 		{
 			void *p = datafile_get_data(df, i);
@@ -271,8 +269,8 @@ bool CDataFileReader::Open(class IStorage *pStorage, const char *pFilename, int 
 					dbg_msg("map", "\tERROR");
 			}
 		}
-		*/
-	}
+		
+	}*/
 
 	return true;
 }
@@ -741,14 +739,14 @@ int CDataFileWriter::Finish()
 	CDatafileHeader Header;
 
 	// we should now write this file!
-	if(DEBUG)
-		dbg_msg("datafile", "writing");
+	//if(DEBUG)
+	//	dbg_msg("datafile", "writing");
 
 	// calculate sizes
 	for(int i = 0; i < m_NumItems; i++)
 	{
-		if(DEBUG)
-			dbg_msg("datafile", "item=%d size=%d (%d)", i, m_pItems[i].m_Size, m_pItems[i].m_Size + (int)sizeof(CDatafileItem));
+		//if(DEBUG)
+		//	dbg_msg("datafile", "item=%d size=%d (%d)", i, m_pItems[i].m_Size, m_pItems[i].m_Size + (int)sizeof(CDatafileItem));
 		ItemSize += m_pItems[i].m_Size + sizeof(CDatafileItem);
 	}
 
@@ -764,8 +762,8 @@ int CDataFileWriter::Finish()
 
 	(void)SwapSize;
 
-	if(DEBUG)
-		dbg_msg("datafile", "num_m_aItemTypes=%d TypesSize=%d m_aItemsize=%d DataSize=%d", m_NumItemTypes, TypesSize, ItemSize, DataSize);
+	//if(DEBUG)
+	//	dbg_msg("datafile", "num_m_aItemTypes=%d TypesSize=%d m_aItemsize=%d DataSize=%d", m_NumItemTypes, TypesSize, ItemSize, DataSize);
 
 	// construct Header
 	{
@@ -783,8 +781,8 @@ int CDataFileWriter::Finish()
 		Header.m_DataSize = DataSize;
 
 		// write Header
-		if(DEBUG)
-			dbg_msg("datafile", "HeaderSize=%d", (int)sizeof(Header));
+		//if(DEBUG)
+		//	dbg_msg("datafile", "HeaderSize=%d", (int)sizeof(Header));
 #if defined(CONF_ARCH_ENDIAN_BIG)
 		swap_endian(&Header, sizeof(int), sizeof(Header) / sizeof(int));
 #endif
@@ -801,8 +799,8 @@ int CDataFileWriter::Finish()
 			Info.m_Type = i;
 			Info.m_Start = Count;
 			Info.m_Num = m_pItemTypes[i].m_Num;
-			if(DEBUG)
-				dbg_msg("datafile", "writing type=%x start=%d num=%d", Info.m_Type, Info.m_Start, Info.m_Num);
+			//if(DEBUG)
+			//	dbg_msg("datafile", "writing type=%x start=%d num=%d", Info.m_Type, Info.m_Start, Info.m_Num);
 #if defined(CONF_ARCH_ENDIAN_BIG)
 			swap_endian(&Info, sizeof(int), sizeof(CDatafileItemType) / sizeof(int));
 #endif
@@ -820,8 +818,8 @@ int CDataFileWriter::Finish()
 			int k = m_pItemTypes[i].m_First;
 			while(k != -1)
 			{
-				if(DEBUG)
-					dbg_msg("datafile", "writing item offset num=%d offset=%d", k, Offset);
+				//if(DEBUG)
+				//	dbg_msg("datafile", "writing item offset num=%d offset=%d", k, Offset);
 				int Temp = Offset;
 #if defined(CONF_ARCH_ENDIAN_BIG)
 				swap_endian(&Temp, sizeof(int), sizeof(Temp) / sizeof(int));
@@ -838,8 +836,8 @@ int CDataFileWriter::Finish()
 	// write data offsets
 	for(int i = 0, Offset = 0; i < m_NumDatas; i++)
 	{
-		if(DEBUG)
-			dbg_msg("datafile", "writing data offset num=%d offset=%d", i, Offset);
+		//if(DEBUG)
+		//	dbg_msg("datafile", "writing data offset num=%d offset=%d", i, Offset);
 		int Temp = Offset;
 #if defined(CONF_ARCH_ENDIAN_BIG)
 		swap_endian(&Temp, sizeof(int), sizeof(Temp) / sizeof(int));
@@ -851,8 +849,8 @@ int CDataFileWriter::Finish()
 	// write data uncompressed sizes
 	for(int i = 0; i < m_NumDatas; i++)
 	{
-		if(DEBUG)
-			dbg_msg("datafile", "writing data uncompressed size num=%d size=%d", i, m_pDatas[i].m_UncompressedSize);
+		//if(DEBUG)
+		//	dbg_msg("datafile", "writing data uncompressed size num=%d size=%d", i, m_pDatas[i].m_UncompressedSize);
 		int UncompressedSize = m_pDatas[i].m_UncompressedSize;
 #if defined(CONF_ARCH_ENDIAN_BIG)
 		swap_endian(&UncompressedSize, sizeof(int), sizeof(UncompressedSize) / sizeof(int));
@@ -872,8 +870,8 @@ int CDataFileWriter::Finish()
 				CDatafileItem Item;
 				Item.m_TypeAndID = (i << 16) | m_pItems[k].m_ID;
 				Item.m_Size = m_pItems[k].m_Size;
-				if(DEBUG)
-					dbg_msg("datafile", "writing item type=%x idx=%d id=%d size=%d", i, k, m_pItems[k].m_ID, m_pItems[k].m_Size);
+				//if(DEBUG)
+				//	dbg_msg("datafile", "writing item type=%x idx=%d id=%d size=%d", i, k, m_pItems[k].m_ID, m_pItems[k].m_Size);
 
 #if defined(CONF_ARCH_ENDIAN_BIG)
 				swap_endian(&Item, sizeof(int), sizeof(Item) / sizeof(int));
@@ -891,8 +889,8 @@ int CDataFileWriter::Finish()
 	// write data
 	for(int i = 0; i < m_NumDatas; i++)
 	{
-		if(DEBUG)
-			dbg_msg("datafile", "writing data id=%d size=%d", i, m_pDatas[i].m_CompressedSize);
+		//if(DEBUG)
+		//	dbg_msg("datafile", "writing data id=%d size=%d", i, m_pDatas[i].m_CompressedSize);
 		io_write(m_File, m_pDatas[i].m_pCompressedData, m_pDatas[i].m_CompressedSize);
 	}
 
@@ -911,7 +909,7 @@ int CDataFileWriter::Finish()
 	io_close(m_File);
 	m_File = 0;
 
-	if(DEBUG)
-		dbg_msg("datafile", "done");
+	//if(DEBUG)
+	//	dbg_msg("datafile", "done");
 	return 0;
 }
